+++ src/main/javascript/Security/OauthFacade.js
@@ -2,6 +2,8 @@ import * as Events from './Events';
 import {OauthToken} from './OauthToken';
 import {OauthConnection} from './OauthConnection';
 
+const defaultProtocolVersion = '2.0';
+
 /**
  * @class
  */
@@ -10,24 +12,35 @@ export class OauthFacade
   /**
    * @param {EventDispatcher} eventDispatcher
    * @param {function} setStorage
-   * @param {String} appId
-   * @param {String} helpdeskUrl
    */
-  constructor(eventDispatcher, setStorage, { appId, helpdeskUrl })
+  constructor(eventDispatcher, setStorage )
   {
-    this.props = { eventDispatcher, setStorage, appId, helpdeskUrl };
+    this.props = { eventDispatcher, setStorage };
   }
 
   /**
    * @method
    *
-   * @param {String} provider
    * @return {Promise}
+   * @param {String} provider
+   * @param options
    */
-  async settings(provider)
+  async settings(provider, options)
   {
     const { eventDispatcher } = this.props;
-    return eventDispatcher.emitAsync(Events.EVENT_SECURITY_SETTINGS_OAUTH, { provider });
+    let eventOptions = null;
+
+   if (typeof options === 'object') {
+     eventOptions = { protocolVersion: defaultProtocolVersion, ...options, provider };
+    } else {
+     eventOptions = { provider, protocolVersion: defaultProtocolVersion };
+    }
+
+    if (eventOptions) {
+      return eventDispatcher.emitAsync(Events.EVENT_SECURITY_SETTINGS_OAUTH, eventOptions);
+    }
+
+    return Promise.reject('invalid argument');
   };
 
   /**
@@ -55,9 +68,16 @@ export class OauthFacade
    */
   async access(provider, options)
   {
+    let eventOptions = null;
+    if (typeof options === 'object') {
+      eventOptions = { ...options, provider };
+    } else {
+      eventOptions = { provider, protocolVersion: defaultProtocolVersion };
+    }
+
     const { eventDispatcher } = this.props;
     return eventDispatcher
-      .emitAsync(Events.EVENT_SECURITY_AUTHENTICATE_OAUTH, { provider, options })
+      .emitAsync(Events.EVENT_SECURITY_AUTHENTICATE_OAUTH, eventOptions)
       .then(OauthToken.fromOauthProxyResponse)
       .then(oauthToken => oauthToken.toJS())
     ;