+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Upcoming
+
+* [FEATURE] allow specifying protocol related requirements when requesting the oauth settings for a provider 
+
 ## v1.0.0-beta.22 - 2017-10-24
 
 * [CHANGE] rename @deskproapps references to @deskpro